Added duplicacy check before adding position to queue again.. moved redis client code to shared project
