+++ liquidation-bot/bot/shared/package.json
@@ -9,7 +9,8 @@
         "web3": "^1.3.4",
         "xdc3": "^1.3.13416",
         "winston": "^3.8.2",
-        "winston-mongodb": "^5.1.0"
+        "winston-mongodb": "^5.1.0",
+        "redis": "^4.4.0"
     },
     "engines": {
         "node": "v0.6.x"
